+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/language/dictWithLabel/DictKeyWithLabelExpressionParser.scala
@@ -5,12 +5,13 @@ import cats.data.{NonEmptyList, Validated}
 import io.circe.parser
 import pl.touk.nussknacker.engine.api.Context
 import pl.touk.nussknacker.engine.api.context.ValidationContext
-import pl.touk.nussknacker.engine.api.definition.{AdditionalVariable => _, DictKeyWithLabelExpression}
+import pl.touk.nussknacker.engine.api.definition.{AdditionalVariable => _}
 import pl.touk.nussknacker.engine.api.expression.{Expression => CompiledExpression, _}
 import pl.touk.nussknacker.engine.api.generics.ExpressionParseError
 import pl.touk.nussknacker.engine.api.typed.typing
 import pl.touk.nussknacker.engine.api.typed.typing.TypingResult
-import pl.touk.nussknacker.engine.graph.expression.Expression
+import pl.touk.nussknacker.engine.graph.expression.{DictKeyWithLabelExpression, Expression}
+import pl.touk.nussknacker.engine.graph.expression.Expression.Language
 import pl.touk.nussknacker.engine.spel.SpelExpressionParseError.KeyWithLabelExpressionParsingError
 
 case class DictKeyWithLabelExpressionTypingInfo(key: String, label: Option[String]) extends ExpressionTypingInfo {
@@ -19,7 +20,7 @@ case class DictKeyWithLabelExpressionTypingInfo(key: String, label: Option[Strin
 
 object DictKeyWithLabelExpressionParser extends ExpressionParser {
 
-  override def languageId: String = Expression.Language.DictKeyWithLabel
+  override def languageId: Language = Expression.Language.DictKeyWithLabel
 
   override def parse(
       keyWithLabel: String,
@@ -53,7 +54,7 @@ object DictKeyWithLabelExpressionParser extends ExpressionParser {
     }
 
   case class CompiledDictKeyExpression(key: String) extends CompiledExpression {
-    override def language: String = languageId
+    override def language: Language = languageId
 
     override def evaluate[T](ctx: Context, globals: Map[String, Any]): T = key.asInstanceOf[T]