+++ docs/Changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-1.13.1 (Not released yet)
+1.13.1 (7 Mar 2024)
 ------------------------
 * [#5447](https://github.com/TouK/nussknacker/pull/5447) Fixed `java.lang.reflect.InaccessibleObjectException: Unable to make public java.lang.Object` exception by downgrade of JRE from 17 to 11 in lite runner image for scala 2.13